Rename device to client
