+++ src/frame/js/actions/auth.js
@@ -1,7 +1,7 @@
 import http from './http';
 import { handleUserConversationResponse } from './conversation';
 
-import { getDeviceId } from '../utils/device';
+import { getClientId } from '../utils/client';
 import { removeItem } from '../utils/storage';
 
 export const SET_AUTH = 'SET_AUTH';
@@ -13,7 +13,7 @@ export function login() {
         return dispatch(http('POST', `/apps/${appId}/login`, {
             userId,
             sessionToken,
-            clientId: getDeviceId(appId)
+            clientId: getClientId(appId)
         })).then(({response, ...props}) => {
             // get rid of session token
             removeItem(`${appId}.sessionToken`);
